Implements StringEndOfParagraphTests
